@@ -13,69 +13,69 @@
   "license": "MIT",
   "private": true,
   "scripts": {
-    "build:dev": "cross-env NODE_ENV=development webpack --progress",
-    "build:prod": "cross-env NODE_ENV=production webpack --progress",
-    "dev": "cross-env ASPNETCORE_ENVIRONMENT=Development NODE_ENV=development dotnet run -c Debug",
-    "prod": "cross-env ASPNETCORE_ENVIRONMENT=Production NODE_ENV=production dotnet run -c Release",
-    "clean": "if exist bin rd /s /q bin & if exist obj rd /s /q obj & if exist publish rd /s /q publish",
-    "publish": "npm run clean && dotnet publish -c Release -o publish -r win-x64 --self-contained",
+    "build:dev": "cross-env NODE_ENV=development ASPNETCORE_ENVIRONMENT=Development webpack --progress",
+    "build:prod": "cross-env NODE_ENV=production ASPNETCORE_ENVIRONMENT=Production webpack --progress",
+    "dev": "cross-env NODE_ENV=development ASPNETCORE_ENVIRONMENT=Development dotnet run -c Debug",
+    "prod": "cross-env NODE_ENV=production ASPNETCORE_ENVIRONMENT=Production dotnet run -c Release",
+    "clean:win": "dotnet clean && (if exist bin rd /s /q bin) & (if exist obj rd /s /q obj) & (if exist publish rd /s /q publish) & (if exist wwwroot rd /s /q wwwroot)",
+    "clean:unix": "dotnet clean && rm -rf bin/ obj/ publish/ wwwroot/",
+    "publish:win": "npm run clean:win && dotnet publish -c Release -o publish -r win-x64 --self-contained",
+    "publish:unix": "npm run clean:unix && dotnet publish -c Release -o publish -r win-x64 --self-contained",
     "lint:config": "eslint --print-config . | eslint-config-prettier-check && stylelint-config-prettier-check",
-    "lint:check": "eslint ./*.js ClientApp/**/*.js ClientApp/**/*.vue && stylelint ClientApp/**/*.css",
-    "lint:fix": "eslint --fix ./*.js ClientApp/**/*.js ClientApp/**/*.vue && stylelint --fix ClientApp/**/*.css",
+    "lint:check": "eslint ./*.js ./*.json ClientApp/**/*.js ClientApp/**/*.vue && stylelint ClientApp/**/*.css",
+    "lint:fix": "eslint --fix ./*.js ./*.json ClientApp/**/*.js ClientApp/**/*.vue && stylelint --fix ClientApp/**/*.css",
     "browsers-list": "npx browserslist",
-    "update-packages": "npx npm-check -u"
+    "update:check": "npx npm-check -u",
+    "update:fix": "npx npm-check -y && npm audit fix"
   },
   "dependencies": {
-    "@babel/runtime": "^7.4.5",
-    "@babel/runtime-corejs3": "^7.4.5",
+    "@babel/runtime": "^7.5.5",
+    "@babel/runtime-corejs3": "^7.5.5",
     "axios": "^0.19.0",
-    "core-js": "^3.1.3",
-    "regenerator-runtime": "^0.13.2",
+    "core-js": "^3.1.4",
+    "overlayscrollbars": "^1.8.0",
+    "regenerator-runtime": "^0.13.3",
     "register-service-worker": "^1.6.2",
     "vue": "^2.6.10",
-    "vue-router": "^3.0.3",
+    "vue-router": "^3.0.7",
     "vue-server-renderer": "^2.6.10",
     "vue-template-compiler": "^2.6.10",
     "vuex": "^3.1.0",
     "vuex-router-sync": "^5.0.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.2.2",
+    "@babel/core": "^7.5.5",
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@babel/plugin-transform-async-to-generator": "^7.4.4",
-    "@babel/plugin-transform-runtime": "^7.4.4",
-    "@babel/preset-env": "^7.4.5",
-    "@fortawesome/fontawesome-svg-core": "^1.2.19",
-    "@fortawesome/free-brands-svg-icons": "^5.9.0",
-    "@fortawesome/free-regular-svg-icons": "^5.9.0",
-    "@fortawesome/free-solid-svg-icons": "^5.9.0",
-    "@fortawesome/vue-fontawesome": "^0.1.6",
+    "@babel/plugin-transform-async-to-generator": "^7.5.0",
+    "@babel/plugin-transform-runtime": "^7.5.5",
+    "@babel/preset-env": "^7.5.5",
+    "@fortawesome/fontawesome-free": "^5.9.0",
     "aspnet-webpack": "^3.0.0",
-    "babel-eslint": "^10.0.1",
+    "babel-eslint": "^10.0.2",
     "babel-loader": "^8.0.5",
     "bootstrap": "^4.3.1",
     "clean-webpack-plugin": "^3.0.0",
     "cross-env": "^5.2.0",
-    "css-loader": "^2.1.1",
-    "eslint": "^5.16.0",
-    "eslint-config-prettier": "^4.3.0",
+    "css-loader": "^3.1.0",
+    "eslint": "^6.1.0",
+    "eslint-config-prettier": "^6.0.0",
     "eslint-plugin-prettier": "^3.1.0",
-    "eslint-plugin-vue": "^5.2.2",
-    "file-loader": "^4.0.0",
+    "eslint-plugin-vue": "^5.2.3",
+    "file-loader": "^4.1.0",
     "jquery": "^3.4.1",
-    "mini-css-extract-plugin": "^0.7.0",
-    "optimize-css-assets-webpack-plugin": "^5.0.1",
+    "mini-css-extract-plugin": "^0.8.0",
+    "optimize-css-assets-webpack-plugin": "^5.0.3",
     "popper.js": "^1.15.0",
     "prettier": "^1.18.2",
     "style-loader": "^0.23.1",
     "stylelint": "^10.1.0",
     "stylelint-config-prettier": "^5.2.0",
     "stylelint-prettier": "^1.1.1",
-    "url-loader": "^2.0.0",
-    "vue-loader": "^15.6.2",
-    "webpack": "^4.33.0",
-    "webpack-cli": "^3.3.3",
-    "webpack-dev-server": "^3.7.1",
+    "url-loader": "^2.1.0",
+    "vue-loader": "^15.7.1",
+    "webpack": "^4.36.1",
+    "webpack-cli": "^3.3.6",
+    "webpack-dev-server": "^3.7.2",
     "webpack-hot-middleware": "^2.25.0"
   },
   "browserslist": [
