@@ -1,5 +1,5 @@
 <template>
-  <div class="about">
+  <article id="app-body">
     <h1>This is an about page</h1>
-  </div>
+  </article>
 </template>
