@@ -3,15 +3,12 @@ import axios from "axios";
 import store from "./store";
 import router from "./router";
 import { sync } from "vuex-router-sync";
-import { FontAwesomeIcon } from "./icons";
-import App from "components/app-root";
+import App from "components/app/index";
 import "./registerServiceWorker";
 
-Vue.component("icon", FontAwesomeIcon);
-
 Vue.prototype.$http = axios;
-
 Vue.config.productionTip = false;
+Vue.config.devtools = false;
 
 sync(store, router);
 
