@@ -2,8 +2,8 @@
 <html lang="en">
 <head>
     <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <title>carlos one</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no, shrink-to-fit=no" />
+    <title>Carlos One</title>
     <link rel="stylesheet" asp-href-include="~/dist/vendor*.css" />
     <link rel="stylesheet" asp-href-include="~/dist/site*.css" />
 </head>
